% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{add_selected_colors}
\alias{add_selected_colors}
\title{Add Selected Colors to Plot}
\usage{
add_selected_colors(
  p,
  selected_palette,
  color_by = NULL,
  color_scale_order = TRUE
)
}
\arguments{
\item{p}{A \code{ggplot}object.}

\item{selected_palette}{Character. The name of the selected color palette.}

\item{color_by}{Optional. Character vector specifying mapped aesthetics.}

\item{color_scale_order}{Logical. Whether to use the standard order of colors (TRUE) or reverse order (FALSE). Defaults to TRUE.}
}
\value{
A \code{ggplot} object with the selected color scales applied.
}
\description{
Adds the selected color palette to a \code{ggplot} object, handling both discrete and continuous mapped aesthetics.
}
