% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeneSymbols.R
\name{addGeneSymbols}
\alias{addGeneSymbols}
\title{Get Gene Symbols}
\usage{
addGeneSymbols(raw_counts, species = c("mouse", "human"))
}
\arguments{
\item{raw_counts}{A data frame containing the raw counts, must contain a \code{GeneID} column containing ensembl identifiers. This can be for instance \code{RawCounts} from \code{\link{createRawCountsWithStats}()}.}

\item{species}{Specify the species for which the gene symbols should be retrieved. At the moment, options are "mouse" or "human". Default is "mouse".}
}
\value{
A list containing:
\itemize{
\item \code{Symbols}: A data frame containing the following columns \code{GeneID}, \code{Symbol}, \code{EntrezID}, \code{Description}, and \code{Alias}. \code{GeneID}s can be duplicated when muliple aliases exist, thus, this is an inflated dataframe but it comes in handy for the alias-based search in |Raw data| -> |Gene expression heatmap|. Gene IDs with missing symbols or missing entrez IDs are removed as well as poorly annotated genes (defined by the occurance of the terms "cdna", "pseudogene" or "predicted" in \code{Description}).
\item \code{CountsSymbols}: The cleaned, depulicated count data frame with unique symbols, ensemble and entrez IDs. Deduplication of symbols is performed using \code{\link[stats]{mad}()} from \code{stats}.
}
}
\description{
Uses a given list of Ensembl Gene IDs to query the appropriate annotation database (mouse or human) and retrieve gene symbols, Entrez IDs, descriptions, and aliases.
}
