% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appServer.R
\name{app_server}
\alias{app_server}
\title{The Server for DExploreR}
\usage{
app_server(input, output, session, config)
}
\arguments{
\item{input}{Default parameter for shiny app servers.}

\item{output}{Default parameter for shiny app servers.}

\item{session}{Default parameter for shiny app servers.}

\item{config}{A list of configuration parameters, such as the data directory.}
}
\description{
This function creates the server for DExplorerR. It relies on many of the functions with the prefix \emph{mod_}.
}
