% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appStartUp.R
\name{app_startup}
\alias{app_startup}
\title{DExploreR App Startup Functions}
\usage{
app_startup()
}
\description{
Performs global setup actions for the DExploreR Shiny app, such as setting the default ggplot2 theme.
}
