% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appUI.R
\name{app_ui}
\alias{app_ui}
\title{DExploreR App UI}
\usage{
app_ui(config)
}
\arguments{
\item{config}{The list of configs that is created in \code{\link{runDExploreR}()}.}
}
\description{
This function defines the user interface of the DExploreR Shiny app and it is called from \code{\link{runDExploreR}()}.
}
