% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCPlotsInfo.R
\name{assign_format_plot_info}
\alias{assign_format_plot_info}
\title{Lookup table for QC plots and info}
\usage{
assign_format_plot_info()
}
\description{
This is the lookup table for the plotting functions and the info texts of the \code{moduleServer} in \code{R/QCServer.R}. It is wrapped in a function so that function references are resolved at call time, not at parse time.
}
