% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDEG.R
\name{calculateDEG}
\alias{calculateDEG}
\title{Calculate Differentially Expressed Genes}
\usage{
calculateDEG(
  raw_counts,
  samples_groups,
  de_fun,
  contrasts = NULL,
  deg_tool = c("limma", "deseq2"),
  method_for_normalization = c("TMM", "TMMwsp", "RLE", "upperquartile", "none"),
  sum_trim = 0.15,
  lm_fit_method = c("robust", "ls")
)
}
\arguments{
\item{raw_counts}{A data frame in the format of \code{RawCountsMeta} from \code{\link{prepareDfs}()}.}

\item{samples_groups}{A data frame in the format of \code{SamplesGroups} from \code{\link{prepareDfs}()}.}

\item{de_fun}{A formula specifying the design for the differential expression analysis. For example, \code{~ 0 + Group} to compare groups defined in the \code{Group} column of \code{samples_groups}. Additional covariates can be included as needed.}

\item{contrasts}{A named vector of contrasts specifying which group comparisons to perform. Cannot be \code{NULL}.}

\item{deg_tool}{Specify which tool to use for DEG calculation. Options are \code{limma} (default) or \code{deseq2}.}

\item{method_for_normalization}{Method of normalization to be used by \code{\link[edgeR]{calcNormFactors}()}. Options are \code{TMM} (default), \code{TMMwsp}, \code{RLE}, \code{upperquartile}, or \code{none}. This parameter is only used when \code{deg_tool = "limma"}.}

\item{sum_trim}{Corresponds to \code{sumTrim} parameter of \code{\link[edgeR]{calcNormFactors}()}. Default is \code{0.15}. This parameter is only used when \code{deg_tool = "limma"} and \code{method_for_normalization = "TMM"}.}

\item{lm_fit_method}{Corresponds to \code{method} parameter of \code{\link[limma]{lmFit}()} from \code{limma}. Options are \code{robust} (default) or \code{ls}. This parameter is only used when \code{deg_tool = "limma"}. Cave: Using \code{ls} is not validated yet. This may require further paramters or other settings than the ones provided.}
}
\value{
A list containing:
\itemize{
\item \code{DEGs}: A tidy data frame with the DEG results for all or the specified contrasts with the following columns: \code{GeneID}, \code{Symbol}, \code{Alias}, \code{EntrezID}, \code{Description}, \code{NCBIURL}, \code{Contrast}, \code{Direction}, \code{AveExpr}, \code{Log2FC}, \code{PVal}, \code{PValAdj}, \code{Rank} and \code{LogPValAdj}.
\item \code{NormCounts}: A normalized expression matrix, calculated by \code{\link[limma]{voom}()} from \code{limma} or by \code{\link[DESeq2]{counts}()} from \code{DESeq2}, with all gene information available added.
}
}
\description{
This function calculates differentially expressed genes (DEGs) between specified sample groups using the \code{edgeR}-\code{limma}-workflow. It calculated DEGs based on the provided raw counts, contrasts and formula.
}
