% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGSEA.R
\name{calculateGSEA}
\alias{calculateGSEA}
\title{Calculate Gene Set Enrichment Analysis}
\usage{
calculateGSEA(
  normalized_count_data = NULL,
  dge_table = NULL,
  species = c("mouse", "human"),
  collection = c("H", "GO:BP", "GO:CC", "GO:MF"),
  contrasts,
  samples_groups
)
}
\arguments{
\item{normalized_count_data}{A data frame with normalized counts with sample names as column names and an additional column called \code{GeneID}. This data frame can be \code{NormData} from \code{\link{calculateDEG}()}, when using \code{deg_tool = "limma"}. Provide this data frame if you want to use \code{\link[gage]{gage}()} for GSEA.}

\item{dge_table}{A data frame with the results of the differential expression analysis containing at least the columns \code{GeneID}, \code{t}, and \code{Contrast}. This data frame can be \code{DEGs} from \code{\link{calculateDEG}()}, when using \code{deg_tool = "limma"}. Provide this data frame if you want to use \code{\link[clusterProfiler]{GSEA}()} for GSEA.}

\item{species}{Select the species for gene set enrichment analysis. Options are \code{mouse} (default) or \code{human}.}

\item{collection}{Select the collection of gene sets for enrichment analysis. Options are \code{H} (hallmark gene sets), \code{GO:BP} (Gene Ontology Biological Process), \code{GO:CC} (Gene Ontology Cellular Component), and \code{GO:MF} (Gene Ontology Molecular Function). Multiple collections can be selected at once by providing a vector.}

\item{contrasts}{The contrasts on which to perform GESA. This can be the exact same vector used by \code{\link{calculateDEG}()}.}

\item{samples_groups}{A data frame assiging samples to groups. For instance, this can be \code{SamplesGroups} from \code{\link{prepareDfs}()}.}
}
\value{
A list containing:
\itemize{
\item \code{ResultsGAGE}: A tidy data frame with the GSEA results (significantly up- and down-regulated gene sets) for all defined contrast.
\item \code{SignificantGeneSetsGenes}: A data frame containing the gene symbols and IDs contained in the significantly enriched gene sets with links to data bases.
}
}
\description{
This function performs Gene Set Enrichment Analysis (GSEA) using the \code{\link[gage]{gage}()} from the \code{gage} based on provided normalized count data or using \code{\link[clusterProfiler]{GSEA}()} from the \code{clusterProfiler} on provided differential expression analysis tables. Both tables can be obtained from \code{\link{calculateDEG}()}. It supports both mouse and human gene sets, which are collected by the \code{msigdbr} package.
}
