% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePlotHeights.R
\name{calculatePlotHeight}
\alias{calculatePlotHeight}
\title{Calculate Dynamic Plot Height}
\usage{
calculatePlotHeight(n_samples, min_size = 500, per_sample_size = 30)
}
\arguments{
\item{n_samples}{Integer. The number of samples to display in the plot.}

\item{min_size}{Numeric. The minimum height (in pixels) for the plot. Default is 500.}

\item{per_sample_size}{Numeric. The height (in pixels) to allocate per sample. Default is 30.}
}
\value{
Numeric. The calculated plot height in pixels.
}
\description{
Dynamically calculates the height of a plot based on the number of samples, ensuring a minimum height for readability.
}
