% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateStats.R
\name{calculateStats}
\alias{calculateStats}
\title{Extract Alignment Statistics For STAR}
\usage{
calculateStats(Tab)
}
\arguments{
\item{Tab}{This is a single ReadsPerGene.out.tab file as a data frame, after read by \code{\link[readr]{read_tsv}()} from the \code{readr} package.}
}
\value{
A data data frame with a single row, containing the sequencing statistics per ReadsPerGene.out.tab tab. This is only used internally by \code{\link{createRawCountsWithStats}()}.
}
\description{
Extract Alignment Statistics For STAR
}
