% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DEXClass.R
\name{check_dataset}
\alias{check_dataset}
\title{Check DExploreR Data Set}
\usage{
check_dataset(object)
}
\arguments{
\item{object}{An object of class "dexDataSet", which will be checked for required columns in the data frame slots.}
}
\description{
Check during class initialization if all required columns are present in the data frame slots of a "dexDataSet" object.
}
