% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBiocPkgs.R
\name{checkBiocPkgs}
\alias{checkBiocPkgs}
\title{Ensure Required Bioconductor Packages Are Installed}
\usage{
checkBiocPkgs(pkgs)
}
\arguments{
\item{pkgs}{A character vector of package names to check.}
}
\value{
Throws an error if any of the specified packages are not installed.
}
\description{
This function checks if the specified Bioconductor packages are installed.
}
