% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurePyEnv.R
\name{configurePyEnv}
\alias{configurePyEnv}
\title{Configure a Python Virtual Environment for DExploreR}
\usage{
configurePyEnv(
  env_name = "r-reticulate-dexplorer",
  env_path = "~/.virtualenvs"
)
}
\arguments{
\item{env_name}{Name of the virtual environment. Default is \code{"r-reticulate-dexplorer"}.}

\item{env_path}{Path to the directory where the virtual environment will be created. Default is \code{"~/.virtualenvs"}.}
}
\value{
The full path to the Python virtual environment as a character string.
}
\description{
This function creates a python environment from R via \code{reticulate}. First, it checks if the environment (\code{"r-reticulate-dexplorer"}) for DExploreR already exists and if not creates a new one using Python 3.12 (or any other Python 3 version if 3.12 is not found). Then, it installs the required Python packages for DExploreR in that environment (\code{"plotly==6.5.2"}, \code{"numpy==2.4.2"}, \code{"kaleido==0.2.1"}, \code{"ipykernel==7.2.0"}). Finally, it returns the path to the created (or existing) virtual environment.
}
