% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCountDistribution.R
\name{createCountDistributionPlot}
\alias{createCountDistributionPlot}
\title{Create Read Count Distribution Plot}
\usage{
createCountDistributionPlot(df, standalone = FALSE)
}
\arguments{
\item{df}{A data frame containing columns for sample names, group assignments, and read counts per gene.}

\item{standalone}{Logical. If \code{TRUE}, the plot is generated as a standalone plot. If \code{FALSE} (required inside DExploreR). Defaults to \code{FALSE}.}
}
\value{
A \code{ggplot2} object, ready for interactive use with \code{plotly}.
}
\description{
Generates a density ridge plot using \code{ggridges} showing the distribution of read counts across recorded genes for each sample and group. Calculates median read counts per gene for tooltips, highlights quartiles, and facets by group.
}
