% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataSet.R
\name{createDataSet}
\alias{createDataSet}
\title{Create A dexDataSet Object}
\usage{
createDataSet(
  QualityControl,
  RawCounts,
  NormalizedCounts,
  VarianceExplained,
  PCA,
  SamplesGroups,
  GeneInfoAliases,
  DGEAnalysis,
  GeneSets,
  GeneSetsGenes,
  pathRawCounts,
  data_path = NULL,
  Cell_line_or_tissue = NULL,
  Study_target = NULL,
  Authors = NULL,
  Date = NULL,
  Details = NULL,
  Sequencing_details = NULL,
  DGEA_details = NULL,
  GSEA_details = NULL,
  pathMetaDataLab = NULL,
  pathMetaDataSeq = NULL
)
}
\arguments{
\item{QualityControl}{A data frame containing quality control metrics per sample, for instance, \code{Stats} from \code{\link{createRawCountsWithStats}()}. Used in |Raw data| -> |Quality control| -> |Number of genes| and |Number of reads|.}

\item{RawCounts}{A data frame containing raw counts per gene and sample, for instance, \code{RawCountsAboveZero} from \code{\link{createRawCountsWithStats}()}. Used in |Raw data| -> |Quality control| -> |Read count distribution|.}

\item{NormalizedCounts}{A data frame containing normalized counts per gene and sample. Used in |Raw data| -> |Gene expression heatmap| and in |GSEA| -> |Top scoring gene sets| -> \code{shiny::modalDialog()}.}

\item{VarianceExplained}{A data frame containing the variance explained by each principal component, for instance, \code{VarianceExplained} from \code{\link{prepareDfs}()}. Used in |Raw data| -> |Dimensionalty reduction|.}

\item{PCA}{A data frame containing them values principal components with sample and group information, for instance \code{DfPCA} from \code{\link{prepareDfs}()}. Used in |Raw data| -> |Dimensionalty reduction|.}

\item{SamplesGroups}{A data frame containing the sample names, their assigned groups and further information for \code{meta_data_lab}, for instance, \code{SamplesGroups} from \code{\link{prepareDfs}()}.}

\item{GeneInfoAliases}{A data frame containing gene annotations such as gene symbols, entrez IDs, descriptions, and aliases and NCBI urls, for instance, \code{GeneSymbols} from \code{\link{addGeneSymbols}()}. This data frame is inflated as each alias gets a separate row.}

\item{DGEAnalysis}{A data frame containing differential gene expression analysis results, for instance, \code{DEGs} from \code{\link{calculateDEG}()}. Used in |DGEA| and in |GSEA| -> |Top scoring gene sets| -> \code{shiny::modalDialog()}}

\item{GeneSets}{A data frame containing gene set enrichment analysis results, for instance, \code{ResultsGAGE} from \code{\link{calculateGSEA}()}. Used in |GSEA|.}

\item{GeneSetsGenes}{A data frame containing genes of enriched gene sets in \code{GAGE}, for instance, \code{SignificantGeneSetsGenes} from \code{\link{calculateGSEA}()}.}

\item{pathRawCounts}{The path to the raw counts directory from STAR alignment.}

\item{data_path}{(Optional) Specify the path to the data directory of the DExploreR app, for instance \verb{/path/to/app/data/data_set_name}, where \verb{data/} is mandatory as it is accessed by DExploreR. Providing \code{data_path} also requires providing \code{Cell_line_or_tissue}, \code{Study_target}, \code{Authors}, \code{Date}, \code{Details}, \code{Sequencing_details}, \code{DGEA_details}, and \code{GSEA_details} for the data set overview in the app (see below). This is information you deliver to a user of DExploreR. When \code{data_path} is provided, then the created data set is saved as an RDS file in the specified path as well as a single row data frame as a .csv with the provided meta information and no return is given. When \code{data_path} is not provided, then the created data set is returned as a list of data frames.}

\item{Cell_line_or_tissue}{What is the cell line or tissue used in the study?}

\item{Study_target}{What is the target of this study, for instance, cancer entitiy?}

\item{Authors}{Which PI, group, or lab performed the study?}

\item{Date}{When was the study performed?}

\item{Details}{What are further details that should be available for the user in DExploreR? This information is displayed as hover information in the data sets overview}

\item{Sequencing_details}{What are details about the sequencing, for instance, library protocol, strategy, ...?}

\item{DGEA_details}{What are details about differential gene expression analysis, for instance, which method was used, which parameters were used, ...?}

\item{GSEA_details}{What are details about gene set enrichment analysis, for instance, which method was used, which parameters were used, ...?
#'}

\item{pathMetaDataLab}{(Optional) The path to the lab metadata input file. Usually this file contains information about grouping of samples and is any how required during preparation of the data.}

\item{pathMetaDataSeq}{(Optional) The path to the sequencing run metadata input file. This file usually contains information about the sequencing run and is any how required during preparation of the data.}
}
\value{
A \code{dexDataSet} object containing all provided data and metadata or no return, when \code{data_path} is provided.
}
\description{
This function creates a \code{dexDataSet} object, which is the data structure used by DExploreR. However, all data sets complying this structure can be used as DExploreR input. Creating of a \code{dexDataSet} object captures also \code{sessionInfo()} and \code{getwd()} for reproducibility. You can get the required classes of all slots and the required columns using \code{\link{printDataSetReqs}()}.
}
