% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardIndexDGEAPlot.R
\name{createDGEAContrastIntersectionPlot}
\alias{createDGEAContrastIntersectionPlot}
\title{Create DGEA Contrast Intersection Plot}
\usage{
createDGEAContrastIntersectionPlot(
  df,
  selected_palette = "App colors",
  standalone = FALSE
)
}
\arguments{
\item{df}{A data frame as returned by \code{\link{formatDGEAContrastIntersection}()}, summarizing Jaccard index results for each pair of contrasts and direction.}

\item{selected_palette}{Character. The name of the color palette to use for the plot. Defaults to "App colors". See \code{print(color_choices)} for available options.}

\item{standalone}{Logical. If \code{TRUE}, the PCA plot is generated as a standalone plot, which is not interactive. If \code{FALSE} (required inside DExploreR), the plot is interactive via \code{plotly}. Defaults to \code{FALSE}.}
}
\value{
An interactive dotplot for Jaccard indicies of up- and downregulated genes as a \code{plotly} object.
}
\description{
Generates an interactive \code{plotly} visualization of Jaccard indices for all pairs of contrasts in a differential gene expression analysis (DGEA). The plot displays the overlap of differentially expressed genes (DEGs) between contrasts, with dot size and color representing the Jaccard index. Tooltips provide detailed comparison information, and facets show results for different regulation directions. Dots are clickable to extract gene lists and venn diagrams (as produced by \code{\link{createVennDiagram}()} for the selected comparison.
}
