% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDetectedGenes.R
\name{createGeneCountPlot}
\alias{createGeneCountPlot}
\title{Create Gene Count Bar Plot}
\usage{
createGeneCountPlot(df, standalone = FALSE)
}
\arguments{
\item{df}{A data frame containing columns for sample names, group assignments, and the number of recorded genes.}

\item{standalone}{Logical. If \code{TRUE}, the plot is generated as a standalone plot. If \code{FALSE} (required inside DExploreR). Defaults to \code{FALSE}.}
}
\value{
A \code{ggplot2} object, ready for interactive use with \code{plotly}.
}
\description{
Generates a horizontal bar plot showing the number of recorded genes per sample and group. Adds tooltips for interactive display, highlights the combined group, and facets by group.
}
