% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRawCountsWithStats.R
\name{createRawCountsWithStats}
\alias{createRawCountsWithStats}
\title{Read STAR Output, Create Raw Counts Data Frame And Statistics Data Frame}
\usage{
createRawCountsWithStats(
  path_to_STAR_directory,
  strandedness = c("CountsUnstranded", "CountsForward", "CountsReverse"),
  fix_runid = NULL
)
}
\arguments{
\item{path_to_STAR_directory}{The path to the STAR output containing the ReadsPerGene.out.tab files.}

\item{strandedness}{Specify strandedness of the library prep protocol to select the correct counting method from ht-seq. Defaults to \code{CountsUnstranded}.}

\item{fix_runid}{A function to fix the \code{RunID}s. Useful when \code{\link[fuzzyjoin]{fuzzy_left_join}()} returns multiple matches. The default is \code{NULL} and it is suggested only to provide a function when the first run resulted in multiple matches.}
}
\value{
A list containing:
\itemize{
\item \code{RawCounts}: A data frame with raw counts (> 0 counts per gene) for all samples. This data frame is required for the ridgeline distribution plots in |Raw data| -> |Quality control| -> |Read count distribution|.
\item \code{Stats}: A data frame with statistics for each sample. This data frame is required for the bar plots in |Raw data| -> |Quality control| -> |Number of reads| and |Raw data| -> |Quality control| -> |Number of genes|.
}
}
\description{
This function reads all \code{ReadsPerGene.out.tab} files from a STAR output directory and combines them into a single data frame. It also calculates some basic statistics for each sample, such as the number of detected genes. The function allows the user to specify the strandedness of the library preparation protocol to select the correct counting method from ht-seq. On demand, it can fix \code{RunID}s using a user-provided function, which is useful when joining functions return multiple matches.
}
