% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screePlot.R
\name{createScreePlot}
\alias{createScreePlot}
\title{Create a Scree Plot}
\usage{
createScreePlot(
  explained_var,
  pc_x = "PC1",
  pc_y = "PC2",
  selected_palette = "App colors",
  standalone = TRUE
)
}
\arguments{
\item{explained_var}{A data frame with columns \code{PC} (principal component names) and \code{Variance} (explained variance for each component).}

\item{pc_x}{Character (can also be a vector of characters specifying multiple PCs or empty for no highlighting). The name of the principal component selected for the x-axis in a related PCA plot. Defaults to "PC1".}

\item{pc_y}{Character (can also be a vector of characters specifying multiple PCs or empty for no highlighting). The name of the principal component selected for the y-axis in a related PCA plot. Defaults to "PC2".}

\item{selected_palette}{Character. The name of the color palette. Defaults to "App colors".}

\item{standalone}{Logical. If \code{TRUE}, the scree plot is generated as a standalone plot, which is not interactive. If \code{FALSE} (required inside DExploreR), the plot is interactive via \code{plotly} and highlights the bars of \code{pc_x} and \code{pc_y}. Defaults to \code{TRUE}.}
}
\value{
The (interactive) scree plot as a \code{ggplot2} or \code{plotly} object.
}
\description{
Generates an (interactive) scree plot using \code{ggplot2} and \code{plotly}, allows to highlight two PCs, for instance, those,which are displayed in the PCA. The plot automatically removes PCs explaining less than 1 \% of total variance and orders the PCs by variance explained.
}
