% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTypeSelector.R
\name{dataTypeSelector}
\alias{dataTypeSelector}
\title{Data Type Selector for DExploreR}
\usage{
dataTypeSelector(config)
}
\arguments{
\item{config}{A list containing the configuration for the data loading, including the mode of operation and the data provided. This parameter is created during initialization of the Shiny app in \code{\link{runDExploreR}()}}
}
\value{
A list containing:
\itemize{
\item \code{dataSetsTable}: A data frame containing the meta data of the available data sets, which is rendered in the \verb{Data sets} tab for selection, when \code{mode == "standard"}. When \code{mode == "internal"}, contained information is only shown in respective tabs of the app, while in \code{mode == "interactive"}, this is \code{NULL} as no selection is necessary.
\item \code{dataSet}: A list containing the data frames of the selected data set, which is used for all downstream analyses. This is \code{NULL}, when \code{dataSetsTable} is not \code{NULL}, thus, when \code{mode == "standard"} and no data set is selected yet. When \code{mode == "internal"} or \code{mode == "interactive"}, this contains the data frames of the internal or user-provided data set, respectively.
}
}
\description{
This function checks the provided data and reads files accordingly.
}
