% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DEXClass.R
\docType{class}
\name{dexDataSet-class}
\alias{dexDataSet-class}
\title{DExploreR Data Set Class}
\description{
This class defines the structure of a DExploreR data set. On top of the actual data, this class also stores metadata such as the working directory and paths to input files and the sessionInfo.
}
