% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexpreprocessr-package.R
\docType{package}
\name{dexpreprocessr-package}
\alias{dexpreprocessr}
\alias{dexpreprocessr-package}
\title{dexpreprocessr: Data Preparation for DExploreR App}
\description{
This package provides a set of functions to prepare data from bulk RNA-seq for visualization with the DExploreR app.
}
\author{
\strong{Maintainer}: Tobias Hundertmark \email{tobias.hundertmark@dkfz-heidelberg.de}

Other contributors:
\itemize{
  \item IBSM - Institute of Medical Bioinformatics and Systems Medicine [copyright holder, funder]
}

}
\keyword{internal}
