% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMods.R
\name{empty_plot}
\alias{empty_plot}
\title{Create an Empty Plot with Custom Message}
\usage{
empty_plot(message = "Nothing to display ...")
}
\arguments{
\item{message}{Character. The message to display in the center of the plot. Default is "Nothing to display ...".}
}
\value{
An interactive \code{plotly} object.
}
\description{
Generates an empty plotly plot displaying a custom message, used to indicate that no data is available for plotting.
}
