% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topGeneSetsDataPrep.R
\name{formatForGeneSetsPlot}
\alias{formatForGeneSetsPlot}
\title{Prepare data form the gene set enrichment plot}
\usage{
formatForGeneSetsPlot(
  df,
  selected_contrast = "all",
  selected_gene_sets = character(0),
  top_gene_sets = 15,
  selected_collections = "all"
)
}
\arguments{
\item{df}{A data frame containing gene set enrichment results, including enrichment scores, set sizes, direction and also a column with the genes of the gene set as a nested tibble.}

\item{selected_contrast}{Character vector. Contrast names to include in the plot. Defaults to "all".}

\item{selected_gene_sets}{Character vector. Gene set names to include in the plot. Defaults to \code{character(0)}, so by default does not filter on specific gene sets.}

\item{top_gene_sets}{Numeric. The number of top absolute enriched gene sets to slice per contrast. Defaults to 15.}

\item{selected_collections}{Character vector. Gene set collection names to include in the plot. Defaults to "all".}
}
\value{
A data frame filtered and prepared for plotting.
}
\description{
Filters for provided contrasts and collections, then slices for top n absolutely enriched gene sets or when specific gene sets are provided, filters for those.
}
