% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapDataPrep.R
\name{formatForHeatmap}
\alias{formatForHeatmap}
\title{Format Data for Gene Expression Heatmap}
\usage{
formatForHeatmap(
  df,
  selected_samples = "keep all",
  selected_subset_size = 20,
  selected_genes = NULL,
  gene_selection_by = TRUE
)
}
\arguments{
\item{df}{A data frame containing Z-scored gene expression values (CPMs) of all samples and genes and additional gene annotation columns.}

\item{selected_samples}{Character vector of sample names to include in the heatmap. Defaults to "keep all", which retains all samples.}

\item{selected_subset_size}{Integer. Number of top genes to select based on median or variance. Defaults to 20.}

\item{selected_genes}{Character vector of gene symbols to always include in the heatmap. This in only relevant in the DExploreR app. Defaults to \code{NULL}.}

\item{gene_selection_by}{Logical. If \code{TRUE}, select top genes by variance; if \code{FALSE}, by median expression. Defaults to \code{TRUE}.}
}
\value{
A data frame formatted for \code{\link[heatmaply]{heatmaply}()}, including z-scored expression values and gene annotation columns.
}
\description{
Prepares and formats a gene expression data frame for a heatmap. Selects top genes by median or variance, ensures user-selected genes are included, removes duplicates, z-scores expression values, and adds gene annotation columns for tooltips and downloads.
}
