% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardIndexGSEADataPrep.R
\name{formatGSEAContrastIntersection}
\alias{formatGSEAContrastIntersection}
\title{Format Data for GSEA Contrast Intersection}
\usage{
formatGSEAContrastIntersection(df)
}
\arguments{
\item{df}{A data frame containing GSEA information, including columns for contrasts, direction and patways.}
}
\value{
A data frame summarizing Jaccard index results for each pair of contrasts and direction, including gene set lists and statistics.
}
\description{
Formats a data frame of enriched gene sets for Jaccard index calculation between pairs of contrasts. Returns a summary table with gene set lists and statistics for each pair.
}
