% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_quiet.R
\name{geom_point_quiet}
\alias{geom_point_quiet}
\title{Quiet version of \code{ggplot2::geom_point()}}
\usage{
geom_point_quiet(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2]{geom_point}()}.}
}
\value{
A ggplot2 layer.
}
\description{
A wrapper for \code{\link[ggplot2]{geom_point}()} that suppresses warnings, for instance when using \code{text} aesthetics to map tooltips for ggplotly.
}
