% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_continuous_scale}
\alias{get_continuous_scale}
\title{Get Continuous Color Scale}
\usage{
get_continuous_scale(family, palette, aes, color_scale_order = TRUE)
}
\arguments{
\item{family}{Character. The color family ("App theme", "Viridis", "Wes Anderson", "RColorBrewer").}

\item{palette}{Character. The palette name.}

\item{aes}{Character. The mapped aesthetic ("fill" or "colour").}

\item{color_scale_order}{Logical. Whether to use the standard order of colors (TRUE) or reverse order (FALSE). Defaults to TRUE.}
}
\value{
A \code{ggplot2} scale object.
}
\description{
Returns a \code{ggplot2} continuous color or fill scale for the specified color family and palette.
}
