% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_discrete_palette}
\alias{get_discrete_palette}
\title{Get Discrete Color Palette}
\usage{
get_discrete_palette(family, palette, n, color_scale_order = TRUE)
}
\arguments{
\item{family}{Character. The color family ("App theme", "Viridis", "Wes Anderson", "RColorBrewer").}

\item{palette}{Character. The palette name.}

\item{n}{Integer. Number of colors to return.}

\item{color_scale_order}{Logical. Whether to use the standard order of colors (TRUE) or reverse order (FALSE). Defaults to TRUE.}
}
\value{
Character vector of color hex codes.
}
\description{
Returns a discrete color palette of length \code{n} for the specified color family and palette.
}
