% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_theme_colors}
\alias{get_theme_colors}
\title{Get Theme Colors}
\usage{
get_theme_colors(color = NULL, n_values = NULL)
}
\arguments{
\item{color}{Character or numeric. Name(s) or index/indices of colors to return. If NULL, returns all theme colors.}

\item{n_values}{Integer. Number of colors to return as a palette. If NULL, returns all or selected colors.}
}
\value{
Character vector of color hex codes.
}
\description{
Returns the app's theme colors as a character vector. Can return all colors, a subset by index or name, or generate a palette of a specified length by interpolating using \code{\link[grDevices]{colorRampPalette}()} from \code{grDevices}.
}
