% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePlotHeights.R
\name{heatmapHeights}
\alias{heatmapHeights}
\title{Calculate Heatmap Component Heights and Domains}
\usage{
heatmapHeights(
  n_genes,
  tiles_height = 40,
  groups_height = 60,
  dendro_height = 160,
  dendro_type = NULL
)
}
\arguments{
\item{n_genes}{Integer. Number of genes (rows) in the heatmap.}

\item{tiles_height}{Numeric. Height (in pixels) allocated per gene for the heatmap tiles. Default is 40 (This proved to look nice.).}

\item{groups_height}{Numeric. Height (in pixels) allocated for group annotation bars. Default is 60 (This proved to look nice.).}

\item{dendro_height}{Numeric. Height (in pixels) allocated for the dendrogram. Default is 160 (This proved to look nice.).}

\item{dendro_type}{Character. Dendrogram display option: "Samples", "Genes", "Samples and genes", or "None". If "Samples" is not present, dendrogram height is set to 0.}
}
\value{
A list with heights for the heatmap components:
\itemize{
\item \code{total_height}
\item \code{dendro_domain}
\item \code{group_domain}
\item \code{tiles_domain}
}
}
\description{
Computes the total height and vertical domains for the main components of a gene expression heatmap (tiles, group annotations, and dendrogram), based on the number of genes and user-specified component heights. Ensures a minimum total height and dynamically adjusts dendrogram height.
}
