% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotControls.R
\name{plotControls}
\alias{plotControls}
\title{Generate Plot Controls UI}
\usage{
plotControls(
  id,
  remove_sample_selection = FALSE,
  remove_color_selection = FALSE,
  ...
)
}
\arguments{
\item{id}{Character. The namespace ID, used for namespacing UI elements.}

\item{remove_sample_selection}{Logical. If TRUE, removes the sample selection control from the plot controls. Default is FALSE.}

\item{remove_color_selection}{Logical. If TRUE, removes the color selection control from the plot controls. Default is FALSE.}

\item{...}{UI element(s) to be included as additional controls in the plot controls section.}
}
\value{
A Shiny UI element (HTML tag list) representing the plot controls.
}
\description{
Creates a set of user interface controls for customizing plots in the DExploreR app. Includes inputs for selecting samples, color palettes, plot settings, and additional controls as needed for specific plots or tabs.
}
