% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareDfs.R
\name{prepareDfs}
\alias{prepareDfs}
\title{Prepare Data Frames For DExploreR}
\usage{
prepareDfs(
  raw_counts,
  meta_data_lab,
  join_meta_data_lab,
  sample_name_column,
  group_column,
  meta_data_seq = NULL,
  join_meta_data_seq = c(RunID = "FASTQ_FILE"),
  min_counts_per_gene = 1,
  min_number_of_samples_per_group_with_gene_expressed = 3,
  min_average_gene_expression_per_group = 1,
  replace_zeros = 0.5,
  log_transform = c("TRUE", "FALSE"),
  centering = c("TRUE", "FALSE"),
  scaling = c("TRUE", "FALSE")
)
}
\arguments{
\item{raw_counts}{A data frame containing raw counts for all samples along with sample names, gene IDs and further information (like symbols, alias ... ) as ouputed by \code{\link{addGeneSymbols}()} as \code{CountsSymbols}.}

\item{meta_data_lab}{A data frame with sample names, grouping information and other relevevant metadata from the lab. This may require custom curation by the user.}

\item{join_meta_data_lab}{Define how to join \code{meta_data_lab} to \code{CountsSymbols}. Ideally, the \code{SAMPLE_NAME} column from \code{meta_data_seq} matches a user-defined sample name column in \code{meta_data_lab}. However, this implementation can also handle partial containments. If the user-defined sample column contains only a fraction of the strings in the \code{SAMPLE_NAME}.}

\item{sample_name_column}{Specify the column to be used for the sample names. You can choose between \code{RunID}, \code{SAMPLE_NAME} and user defined column name. \code{RunID} comes from \code{CountsSymbols} of \code{\link{addGeneSymbols}()} but this is the sequencing run IDs, or the STAR output directory names. \code{SAMPLE_NAME} comes from \code{meta_data_seq} and the sample name column you specified comes from \code{meta_data_lab}. Most likely, this is the one you want to use since this is how your collaborator knows their samples. Ideally, this (partially) matches the \code{SAMPLE_NAME} in \code{meta_data_seq}.}

\item{group_column}{\code{meta_data_lab} should contain columns, which define groups of samples. This parameter should not be left empty, but you can also only specify one column. The resulting column will get the generic name \code{Group}.}

\item{meta_data_seq}{A data frame with the sequencing related metadata. This may require custom curation by the user. This data frame is required to match the groups from \code{meta_data_lab} to the samples in \code{CountsSymbols} via the sequencing run IDs. When sample names from \code{meta_data_lab} can directly be mapped to the \code{RunID}s in \code{CountsSymbols}, this parameter can be left as \code{NULL}.}

\item{join_meta_data_seq}{Define how to join \code{meta_data_seq} to \code{CountsSymbols}. The default assumes that the \code{RunID} column in \code{CountsSymbols} contains the sequencing run IDs, which can be matched to the \code{FASTQ_FILE} column in \code{meta_data_seq}. Yet, you need to make sure that \code{FASTQ_FILE} does not contain the \verb{_R[12].fastq.gz} suffix. You can use: \code{mutate(FASTQ_FILE = sub("_R[12].fastq.gz", "", FASTQ_FILE))}, for instance. This parameter will be ignored when \code{meta_data_seq == NULL}.}

\item{min_counts_per_gene}{Define the minimum number of read counts per gene for this to be considered expression.}

\item{min_number_of_samples_per_group_with_gene_expressed}{Define the minimum number of samples per group that need to express a gene (as defined by \code{min_counts_per_gene}) for this gene to be retained for PCA and further analysis.}

\item{min_average_gene_expression_per_group}{Define the minimum average gene expression per group for a gene to be retained for PCA and further analysis.}

\item{replace_zeros}{For PCA, the raw counts are converted to log-CPM values using \code{\link[edgeR]{cpm}()} from \code{edgeR}. To avoid issues with \eqn{log(0)}, all zero values are replaced by this value prior to log-CPM conversion.}

\item{log_transform}{Toggle \eqn{log2} transformation during CPM calculation. Default is \code{TRUE}.}

\item{centering}{Toggle centering during PCA calculation. Default is \code{TRUE}.}

\item{scaling}{Toggle scaling during PCA calculation. Default is \code{FALSE}.}
}
\value{
A list containing:
\itemize{
\item \code{RawCountsMeta}: This data frame is very similar to \code{CountsSymbols} from \code{\link{addGeneSymbols}()} but contains user-defined columns with sample names and grouping information. Furthermore, this can also be filtered based on the number of samples per group that need to have a certain gene expressed and based on the average gene expression per group. Setting \code{min_number_of_samples_per_group_with_gene_expressed} and \code{min_average_gene_expression_per_group} to \eqn{0} will disable this filtering.
\item \code{SamplesGroups}: This assignes each sample to its group, which was defined by \code{group_column}. This data frame contains the columnes \code{Group} and \code{SampleNameUser} along with the \code{RunID} and other columns from \code{meta_data_lab} that are related to sample names and groups. This data frame can be used to translate between different sample names and groups.
\item \code{CPMCounts}: This data frame contains the CPM values calculated from the raw counts after filtering.
\item \code{RawPCA}: This is the raw result of the PCA as outputed by \code{\link[stats]{prcomp}()} from \code{stats}.
\item \code{DfPCA}: This is the data frame containing the x and y coordinates for each sample and PC. It also contains the \code{RunID} and \code{Group} columns for plotting. This data frame is required for the PCA plot in |Raw data| -> |Dimensionaly reduction|.
\item \code{VarianceExplained}: This data frame contains the percentage of variance explained by each PC. It will be used for the Scree plot in |Raw data| -> |Dimensionaly reduction| and for axis labels in the PCA plot.
}
}
\description{
This function combines the \code{CountsSymbols} from \code{\link{addGeneSymbols}()} with meta data from the lab and the sequencing run. Furthermore, it provides some functionality for filtering. It calculates CPMs using \code{\link[edgeR]{cpm}()} from \code{edgeR} and feeds them into \code{\link[stats]{prcomp}()} from \code{stats} for PCA. The data frames genereated herewith are required for all further anylses in the DExploreR shiny app.
}
