% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DEXClass.R
\name{printDataSetReqs}
\alias{printDataSetReqs}
\title{Print requirements for DExploreR data set}
\usage{
printDataSetReqs()
}
\description{
This function prints the class of each slot of the \code{dexDataSet} class and if the slot is a dataframe it prints the required columns. This can be useful for users to check if their data frames meet the requirements before creating a \code{dexDataSet} object.
}
