% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDExploreR.R
\name{runDExploreR}
\alias{runDExploreR}
\title{Run DExploreR Shiny App}
\usage{
runDExploreR(
  asset_dir = "/Users/tobiashundertmark/Downloads",
  data_dir = "/Users/tobiashundertmark/Downloads/data"
)
}
\description{
...
}
