% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGAGE.R
\name{runGAGE}
\alias{runGAGE}
\title{Run GAGE for Gene Set Enrichment Analysis}
\usage{
runGAGE(condition, control, norm_count, gene_sets, samples_groups)
}
\arguments{
\item{condition}{The name of the first contrast, i.e. the condition.}

\item{control}{The name of the second contrast, i.e. the control.}

\item{norm_count}{The normalized count matrix sample names as columns and row names as gene IDs, subset to the well-annotate genes.}

\item{gene_sets}{A list of gene sets with genes as ensembl IDs. These will be analaysed for enrichment.}

\item{samples_groups}{A list of contrasts with assigned sample names.}
}
\value{
A tidy data frame with significantly enriched (up- and down-regulated) gene sets. The column names match the ones used by DExploreR for compatibility.
}
\description{
This function is a wrapper for \code{\link[gage]{gage}()} from the \code{gage} package using the provided contrasts and samples.
}
