% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1DEXClass.R
\docType{class}
\name{sessionInfo-class}
\alias{sessionInfo-class}
\title{Session Info Class}
\description{
Explicitly define an S4 wrapper class for \code{sessionInfo}, so that this can be set in the slot definitions.
}
