% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMods.R
\name{standalone_plot_style}
\alias{standalone_plot_style}
\title{Adjust the plot style for standalone plots}
\usage{
standalone_plot_style(p)
}
\arguments{
\item{p}{A \code{ggplot} object to which the standalone style will be applied.}
}
\value{
A styled \code{ggplot} object.
}
\description{
Some styling for standalone plots, i.e. for plots used outside of DExploreR. Add white background, grey grid lines and remove the strip text background.
}
