% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCUi.R
\name{tabContentUI}
\alias{tabContentUI}
\title{Quality Control Tab UI}
\usage{
tabContentUI(id)
}
\arguments{
\item{id}{Character. The module namespace ID.}
}
\value{
A Shiny UI element (HTML tag list) for inclusion in the app UI.
}
\description{
Generates the user interface for the Quality Control tab in the DExploreR app. Provides controls for selecting plot type, color palette, and samples, as well as buttons for further information, data download, and plot download. Displays the selected plot using \code{plotly}. This has the same layout as \code{\link{makeSubTabContent}()}.
}
