% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexDataSetConversions.R
\name{toList}
\alias{toList}
\title{Convert DexDataSet Object to a List}
\usage{
toList(object)
}
\arguments{
\item{object}{A \code{dexDataSet} object to be converted.}
}
\description{
This function converts a \code{dexDataSet} object into a standard R list, preserving all its components. This avoids dependency on on this package, when working with the data.
}
