% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexDataSetConversions.R
\name{toObject}
\alias{toObject}
\title{Convert List of Data frames to a DexDataSet Object}
\usage{
toObject(list)
}
\arguments{
\item{list}{A list of data frames, which should be converted to a \code{dexDataSet} object. The list should contain the same components as the \code{dexDataSet} object, which are: \code{QualityControl}, \code{RawCounts}, \code{NormalizedCounts}, \code{VarianceExplained}, \code{PCA}, \code{SamplesGroups}, \code{GeneInfoAliases}, \code{DGEAnalysis}, \code{GeneSets}, \code{GeneSetsGenes}, \code{sessionInfo}, \code{workingDirectory}, \code{pathRawCounts}, \code{pathMetaDataLab} and \code{pathMetaDataSeq}. The class operator will then check if all necessary columns are available, so if the data set is valid for loading into the app.}
}
\description{
This functions turns a list of data frames into a \code{dexDataSet} object. This is especially useful, when checking the validity of user-provided data sets. These are provided as lists of data frames as an \code{.rds} file.
}
