% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plotHeights.R
\name{calculateDomains}
\alias{calculateDomains}
\title{Calculate Facet Plot Domains}
\usage{
calculateDomains(df, total_height, white_space = 35)
}
\arguments{
\item{df}{A data frame with columns \code{group} (group/facet identifier), \code{n_samples} (number of samples in each group) and \code{yaxis} (name of the y-axis form the plotly object).}

\item{total_height}{The total height (in pixels) allocated for the plot.}

\item{white_space}{The amount of white space (in pixels) to insert between facets. Default is 35 (this seems suitable based on experience).}
}
\value{
The input data frame with an added \code{domain} column (a list-column of length-2 numeric vectors giving the lower and upper bounds for each facet's domain).
}
\description{
Calculates the vertical domains for each group/facet in a facetted plot, ensuring proportional heights based on the number of samples per group and consistent bar widths. Adds white space between facets and returns the original data frame with an added \code{domain} column for each group.
}
