% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_colors.R
\name{create_heatmap_color_function}
\alias{create_heatmap_color_function}
\title{Create Heatmap Color Function}
\usage{
create_heatmap_color_function(selected_colors)
}
\arguments{
\item{selected_colors}{Character. The name of the selected color palette.}
}
\value{
A function that generates a color palette of arbitrary length.
}
\description{
Creates a color function for heatmaps using the selected color palette, suitable for use with \code{\link[heatmaply]{heatmaply}()}.
}
