% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardIndexGSEAPlot.R
\name{createGSEAContrastIntersectionPlot}
\alias{createGSEAContrastIntersectionPlot}
\title{Create GSEA Contrast Intersection Plot}
\usage{
createGSEAContrastIntersectionPlot(
  df,
  selected_palette = "App colors",
  standalone = FALSE
)
}
\arguments{
\item{df}{A data frame as returned by \code{\link{formatGSEAContrastIntersection}()}, summarizing Jaccard index results for each pair of contrasts and direction.}

\item{selected_palette}{Character. The name of the color palette to use for the plot.}

\item{standalone}{Logical. If \code{TRUE}, the PCA plot is generated as a standalone plot, which is not interactive. If \code{FALSE} (required inside DExploreR), the plot is interactive via \code{plotly}. Defaults to \code{FALSE}.}
}
\value{
An interactive dotplot for Jaccard indicies for enriched gene sets by contrasts as a \code{plotly} object.
}
\description{
Generates an interactive \code{plotly} visualization of Jaccard indices for all pairs of contrasts in a gene set enrichment analysis (GSEA). The plot displays the overlap of enriched gene sets between contrasts, with dot size and color representing the Jaccard index. Tooltips provide detailed comparison information, and facets show results for different regulation directions.
}
