% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAPlot.R
\name{createPCAPlot}
\alias{createPCAPlot}
\title{Create PCA Plot}
\usage{
createPCAPlot(
  df_pca,
  explained_var,
  pc_x = "PC1",
  pc_y = "PC2",
  selected_palette = "App colors",
  group_overlay = "Convex hull",
  standalone = FALSE
)
}
\arguments{
\item{df_pca}{A data frame containing PCA results, including columns for sample names, group assignments, and principal component scores.}

\item{explained_var}{A data frame with columns \code{PC} and \code{Variance}, giving the explained variance for each principal component.}

\item{pc_x}{Character. The name of the principal component to plot on the x-axis. Default is "PC1".}

\item{pc_y}{Character. The name of the principal component to plot on the y-axis. Default is "PC2".}

\item{selected_palette}{Character. The name of the color palette to use for group coloring.}

\item{group_overlay}{Character. Overlay type for group visualization: "Ellipse" for ellipses with 95 \% confidence interval, "Convex hull" for convex hulls, or any of  "none", "no overlay", "remove overlay" or "w/o" to hide the group overlay. Default is "Convex hull".}

\item{standalone}{Logical. If \code{TRUE}, the PCA plot is generated as a standalone plot, which is not interactive. If \code{FALSE} (required inside DExploreR), the plot is interactive via \code{plotly}. Defaults to \code{FALSE}.}
}
\value{
The PCA plot as \code{ggplot2} object.
}
\description{
Generates a PCA plot using \code{ggplot2}.
}
