% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_screePlot.R
\name{createScreePlot}
\alias{createScreePlot}
\title{Create Interactive Scree Plot}
\usage{
createScreePlot(explained_var, pc_x, pc_y, selected_palette)
}
\arguments{
\item{explained_var}{A data frame with columns \code{PC} (principal component names) and \code{Variance} (explained variance for each component).}

\item{pc_x}{Character. The name of the principal component selected for the x-axis in a related PCA plot.}

\item{pc_y}{Character. The name of the principal component selected for the y-axis in a related PCA plot.}

\item{selected_palette}{Character. The name of the color palette.}
}
\value{
The interactive scree plot as a \code{plotly} object.
}
\description{
Generates an interactive scree plot using \code{ggplot2} and \code{plotly}.
}
