% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dexplorer-package.R
\docType{package}
\name{dexplorer-package}
\alias{dexplorer}
\alias{dexplorer-package}
\title{dexplorer: Visualize and interactively explore bulk RNA-seq data}
\description{
This package contains a shiny app for the visualization and interactive exploration of bulk RNA-seq data. Furthermore, it provides a set of functions to prepare RNA-seq data in a suitable foramt for DExploreR.
}
\author{
\strong{Maintainer}: Tobias Hundertmark \email{tobias.hundertmark@dkfz-heidelberg.de}

Other contributors:
\itemize{
  \item IBSM - Institute of Medical Bioinformatics and Systems Medicine [copyright holder, funder]
}

}
\keyword{internal}
