% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_downloadUIElements.R
\name{downloadSettingsModal}
\alias{downloadSettingsModal}
\title{Download Settings for Downloads from Modals}
\usage{
downloadSettingsModal(id)
}
\arguments{
\item{id}{Character. The namespace ID
.}
}
\value{
A Shiny UI element (HTML tag list) representing the download settings modal.
}
\description{
Generates a UI elements for customizing plot download settings in the DExploreR app. Allows users to select file format, specify plot width and height, and trigger the download of the current plot.
}
