% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_colors.R
\name{get_color_family}
\alias{get_color_family}
\title{Get Color Family of a Selected Palette}
\usage{
get_color_family(selected_palette)
}
\arguments{
\item{selected_palette}{Character. The palette name.}
}
\value{
Character. The color family name.
}
\description{
Returns the color family name (e.g., "Viridis", "Wes Anderson") for a selected palette. Although this information is contained in \code{color_choices}, the corresponding UI element returns an unnamed value.
}
