% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_colors.R
\name{get_continuous_scale}
\alias{get_continuous_scale}
\title{Get Continuous Color Scale}
\usage{
get_continuous_scale(family, palette, aes)
}
\arguments{
\item{family}{Character. The color family ("App theme", "Viridis", "Wes Anderson", "RColorBrewer").}

\item{palette}{Character. The palette name.}

\item{aes}{Character. The mapped aesthetic ("fill" or "colour").}
}
\value{
A \code{ggplot2} scale object.
}
\description{
Returns a \code{ggplot2} continuous color or fill scale for the specified color family and palette.
}
