% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_subTabLayout.R
\name{makeSubTabContent}
\alias{makeSubTabContent}
\title{Create Consistent Sub Tab Layout}
\usage{
makeSubTabContent(
  id,
  further_controls = div(),
  top_left_wide = div(),
  main_content = div(),
  remove_sample_selection = FALSE,
  remove_color_selection = FALSE,
  additional_button_right = div()
)
}
\arguments{
\item{id}{Character. The namespace ID, used for namespacing UI elements. This is passed on to \code{\link{plotControls}()} to ensure that the controls are properly namespaced, too.}

\item{further_controls}{UI element(s) to be included as additional controls in \code{\link{plotControls}()}. Default is \code{div()}.}

\item{top_left_wide}{UI element to be displayed in the top left wide area. Default is \code{div()}.}

\item{main_content}{UI element(s) for the main content area (e.g., plots or tables). Default is \code{div()}.}

\item{remove_sample_selection}{Logical. If TRUE, removes the sample selection control from \code{\link{plotControls}()}. Default is FALSE.}

\item{remove_color_selection}{Logical. If TRUE, removes the color selection control from \code{\link{plotControls}()}. Default is FALSE.}

\item{additional_button_right}{UI element for an additional button to be placed on the right. Default is \code{div()}.}
}
\value{
A Shiny UI element (HTML tag list) representing the sub-tab layout.
}
\description{
Generates a consistent UI layout for sub-tabs in the DExploreR app, including plot controls, information and download buttons, and a main content area for plots or tables. Allows insertion of additional controls and buttons as needed for specific tabs.
}
