% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDExploreR.R
\name{runDExploreR}
\alias{runDExploreR}
\title{Run DExploreR Shiny App}
\usage{
runDExploreR(
  asset_dir = "/Users/tobiashundertmark/Downloads",
  data_dir = "/Users/tobiashundertmark/Downloads/data"
)
}
\arguments{
\item{asset_dir}{A directory containing the assets for DExploreR, such as logos.}

\item{data_dir}{A directory containing \code{.rds} files and corresponding \code{.csv}s as prepared by \code{\link{createDataSet}()}.}
}
\description{
...
}
