% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_serverQC.R
\name{tabContentServer}
\alias{tabContentServer}
\title{Quality Control Tab Server Logic}
\usage{
tabContentServer(id, data, plot_status, authors)
}
\arguments{
\item{id}{Character. The module namespace ID.}

\item{data}{A reactive expression returning a list of data frames required for plotting. This is \code{data_set_loaded()} in \code{\link{app_server}()}.}

\item{plot_status}{A reactiveValues object for tracking plot readiness.}

\item{authors}{Character. Author information for use in download file names.}
}
\description{
Implements the server-side logic for the Quality Control tab in the DExploreR app. Handles plot selection, color palette and sample selection, data filtering, plot generation, download handlers, and modal dialogs for further information and plot downloads. Integrates with \code{plotly} for interactive visualization and dynamically adjusts plot layout and tooltips.
}
