% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subTabLayout.R
\name{tabHeaders}
\alias{tabHeaders}
\title{Generate Tab Headers}
\usage{
tabHeaders(title, text_id)
}
\arguments{
\item{title}{Character. The title to display in the tab header and browser window.}

\item{text_id}{Character. The output ID for the descriptive text to be rendered below the title.}
}
\value{
A Shiny UI element (HTML tag list) representing the tab header.
}
\description{
Creates a header layout for a tab in the DExploreR app, including a title panel and a text output area. Used to provide a consistent header with a title and descriptive text for each tab.
}
